+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0]
+
 ## Added
 - Support to nullable page properties (#149)
   - Properties: `Date`, `Email`, `Number`, `PhoneNumber`, `Select`, `Url`.
@@ -235,4 +237,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.0.0-beta.2]: https://github.com/mariosimao/notion-sdk-php/releases/tag/v1.0.0-beta.2
 [1.0.0]: https://github.com/mariosimao/notion-sdk-php/releases/tag/v1.0.0
 [1.1.0]: https://github.com/mariosimao/notion-sdk-php/releases/tag/v1.1.0
-[Unreleased]: https://github.com/mariosimao/notion-sdk-php/compare/v1.1.0...HEAD
+[1.2.0]: https://github.com/mariosimao/notion-sdk-php/releases/tag/v1.2.0
+[Unreleased]: https://github.com/mariosimao/notion-sdk-php/compare/v1.2.0...HEAD